# Conflicts:
#	src/main/java/com/cannolicatfish/rankine/init/ModRecipes.java
#	src/main/resources/assets/rankine/textures/gui/rankine_journal.png

+++ src/main/java/com/cannolicatfish/rankine/util/elements/ElementInterface.java
@@ -8,7 +8,7 @@ public interface ElementInterface {
 
     int getDurabilityFromPercent(int x);
 
-    int getDamageFromPercent(int x);
+    float getDamageFromPercent(int x);
 
     float getAttackSpeedFromPercent(int x);
