+++ src/main/java/com/cannolicatfish/rankine/util/elements/LeadElement.java
@@ -20,7 +20,7 @@ public class LeadElement implements ElementInterface {
     }
 
     @Override
-    public int getDamageFromPercent(int x) {
+    public float getDamageFromPercent(int x) {
         return 0;
     }
