+++ src/main/java/com/cannolicatfish/rankine/util/elements/NetheriteElement.java
@@ -16,7 +16,7 @@ public class NetheriteElement implements ElementInterface {
     }
 
     @Override
-    public int getDamageFromPercent(int x) {
+    public float getDamageFromPercent(int x) {
         return 0;
     }
